Moving classes that only contain Harmony patches into that namespace
