+++ CraftMagicItems/Main.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Reflection.Emit;
-using System.Runtime.Serialization;
 using System.Text.RegularExpressions;
 using CraftMagicItems.Config;
 using CraftMagicItems.Constants;
@@ -16,13 +14,10 @@ using CraftMagicItems.UI.Sections;
 using CraftMagicItems.UI.UnityModManager;
 using Kingmaker;
 #if PATCH21
-using Kingmaker.Assets.UI.Context;
 #endif
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
-using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Blueprints.Classes.Spells;
-using Kingmaker.Blueprints.Facts;
 using Kingmaker.Blueprints.Items;
 using Kingmaker.Blueprints.Items.Armors;
 using Kingmaker.Blueprints.Items.Ecnchantments;
@@ -32,12 +27,8 @@ using Kingmaker.Blueprints.Items.Weapons;
 using Kingmaker.Blueprints.Loot;
 using Kingmaker.Blueprints.Root;
 using Kingmaker.Blueprints.Root.Strings.GameLog;
-using Kingmaker.Controllers.Rest;
-using Kingmaker.Designers;
 using Kingmaker.Designers.Mechanics.EquipmentEnchants;
 using Kingmaker.Designers.Mechanics.Facts;
-using Kingmaker.Designers.Mechanics.WeaponEnchants;
-using Kingmaker.Designers.TempMapCode.Capital;
 using Kingmaker.EntitySystem.Entities;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
@@ -51,14 +42,10 @@ using Kingmaker.Kingdom;
 using Kingmaker.Localization;
 using Kingmaker.PubSubSystem;
 using Kingmaker.RuleSystem;
-using Kingmaker.RuleSystem.Rules;
 using Kingmaker.RuleSystem.Rules.Abilities;
-using Kingmaker.RuleSystem.Rules.Damage;
 using Kingmaker.UI;
-using Kingmaker.UI.ActionBar;
 using Kingmaker.UI.Common;
 using Kingmaker.UI.Log;
-using Kingmaker.UI.Tooltip;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
@@ -69,16 +56,14 @@ using Kingmaker.UnitLogic.Alignments;
 using Kingmaker.UnitLogic.Buffs;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.FactLogic;
-using Kingmaker.UnitLogic.Mechanics.Components;
-using Kingmaker.UnitLogic.Parts;
 using Kingmaker.Utility;
-using Kingmaker.View.Equipment;
 using Newtonsoft.Json;
 using UnityEngine;
 using UnityModManagerNet;
 using Random = System.Random;
 
-namespace CraftMagicItems {
+namespace CraftMagicItems
+{
     public static class Main {
         private const string BondedItemRitual = "bondedItemRitual";
 
@@ -820,7 +805,7 @@ namespace CraftMagicItems {
         }
 
         public static bool IsOversized(BlueprintItem blueprint) {
-            return GetEnchantments(blueprint).Any(enchantment => enchantment.AssetGuid.StartsWith(ItemQualityBlueprints.OversizedGuid) && !enchantment.GetComponent<WeaponSizeChange>());
+            return GetEnchantments(blueprint).Any(enchantment => enchantment.AssetGuid.StartsWith(ItemQualityBlueprints.OversizedGuid) && !enchantment.GetComponent<Patches.Harmony.WeaponSizeChange>());
         }
 
         // Use instead of UIUtility.IsMagicItem.
