+++ CraftMagicItems/Patches/CraftMagicItemsBlueprintPatcher.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Text.RegularExpressions;
 using CraftMagicItems.Constants;
 using CraftMagicItems.Localization;
+using CraftMagicItems.Patches.Harmony;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Items;
@@ -23,6 +24,7 @@ using Kingmaker.UI.Common;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.ActivatableAbilities;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
+using Kingmaker.Utility;
 using Kingmaker.View.Animation;
 using Kingmaker.Utility;
 #if PATCH21_BETA
